refactor/test: resturcture to map directly to template and add test
