--- examples/overloaded-cpp-c-abi-wrappers/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "overloaded-cpp-c-abi-wrappers"
-version = "0.1.0"
-edition = "2024"
-# Run build.rs so Cargo compiles and links the C++ source.
-build = "build.rs"
-
-[build-dependencies]
-# `cc` is used by build.rs to compile C/C++ code.
-cc = "1.2"
