+++ examples/overloaded-multiply-by-2/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    cpp_build::build("src/main.rs");
+}
